Make listeer socket inode a 64 bit integer, not 32
Inside the kernel, the socket inode is stored as a 64-bit value. It is
also printed out as %lu in /proc/net/tcp, which means that on 64-bit
systems, the full 64-bit inode number will be printed out in that file
too.

UNFORTUNATELY, the netlink inet_diag API only exposes the listener
socket inode as a 32 bit number??? It just silently wraps it around if
you have a lot of inodes.

Because of this:
- Internally store the inode in raingutter as uint64, since we can get
  that from /proc
- Accept the wrapped-around value from netlink, since we have no choice
  there
- Mark /proc as the preferred and most correct option for raingutter
  socket stats.